Move event_pool and message_queue to core.
