+++ orttraining/orttraining/core/session/training_session.cc
@@ -33,7 +33,7 @@
 #include "core/providers/cuda/cuda_allocator.h"
 #endif
 
-#include "orttraining/training_ops/cpu/controlflow/event_pool.h"
+#include "core/providers/cpu/controlflow/event_pool.h"
 #if defined(USE_CUDA) && defined(ORT_USE_NCCL) && defined(USE_NCCL_P2P)
 #include "orttraining/training_ops/cuda/communication/nccl_service.h"
 #endif
